Replace square brackets with angle brackets in pretty-printer tests to make failures easier to diagnose
In the ScalaTest failure output it uses square brackets to indicate which parts of the test output are
not as expected. If the output itself contains square brackets it is hard to tell which are real ones
and which are indicating a test failure. Using angle brackets instead makes it easier to tell.