Remove dependency on Scala parser combinator library (and flow-on effects)
The dependency on the parser combinator library was causing problems for a
number of reasons, particularly since it's not maintained much these days and
it is too hard to extend it from Kiama to do what we want. In particular,
supporting a more sophisticated view of source positions turned out to be
hard. We also already had somewhat messy extensions to add better white space
handling which had to be kept in line with the library.

So, we have a local Kiama parsing combinator library again. The aim was to get
something that was much simpler than the Scala library which supports things
like many different kinds of input sources, a few different kinds of parsers,
etc, but also be source compatible as much as possible. The new code is
similar in many ways to the Scala library but is much simpler. It resurrects
some old, old Kiama code that added support for left recursion.

The new parsing library implementation prompted some other related changes. In
particular, our preferred view now is to use vectors to represent the repeated
children in nodes instead of cons-based lists. The reasoning is that there
should be the same access to each child rather than access that is biased to
the earlier children. To this end, the parsing class Parsers uses vectors as
the return type for repetition combinators such as rep, rep1 and the postfix *
and + operators. The ListParsers class is identical except it uses lists
instead of vectors.

A summary of the main changes follows:

* drop build dependency on Scala parsing library
* new parsing library in org.kiama.parsing.Parsers
* types are Parser[T] and PackratParser[T], the latter supports memoisation
and left recursion
* not all combinators of the old library are supported, just the common ones
* the old ParserUtilities, PositionedParserUtilities etc are gone, this is
all supported by default in the parsing library

* the definition of whitespace in the new parsers is different to that in
the old ones, previously the whitespace parser matched whitespace that was
present (e.g., \s+) but in the new version it must also match in the case
where there is no whitespace (e.g., \s*)
** as a result of this change, it is now possible to have two parsed
constructs that are not separated by present whitespace (as long as the
first parser doesn't consume what the second one wants)

* vectors are used for most repeated constructs instead of lists
* use org.kiama.parsing.ListParsers if you prefer to use lists for repetition
* all of the examples use the new parsing library, most examples use vectors
but dataflow and til use lists
* some other parts of Kiama that talk about children of nodes now use vectors
(e.g., the Term deconstructor in rewriting)
* in some cases tests have been broadened to test lists and vectors, in other
cases list tests have been replaced by vector ones
* relations (including the tree ones) now also use a vector for the basic
data structure instead of a list

* the Compiler framework no long mixes in the parsing modules
** there is now a parsers local that must be defined as well as a parser
which is the main entry point you want the Compiler to use

* the new Kiama positions support filenames as well as line and column
** all messages now include the filename if available

* the old global position store is no more, you can mix one in using the trait
PositionStore
** Compilers, REPL etc have a position store
* parsing modules must be provided with a position store
** usually the easiest way to do this is to pass in the store used by a Compiler
so that the parsers and the compiler share the store for messaging
* printed positions don't include a blank link before the pointer (which was
due to a bug in the Scala library)
* the support for the "with whitespace included" position of something has
been dropped since we really didn't use it and it's non-trivial to do with
the new parsing library

* as a result of the position change, the position-dependent parts of the
Messaging library are now also not global
** Compiler, REPL etc have a messaging instantiation
* some methods in the messaging library have been given better names,
e.g., formatMessage instead of formats
* the sortmessages method has been removed in favour of just using .sorted
on a message list since we now provide an Ordering instance for messages