Remove final modifier check when unboxing AnyVal Products during rewriting
This check doesn?t work on Scala 2.10 due to the bug that is addressed by the following change:

https://github.com/scala/scala/pull/2636

The change wasn?t back-ported to 2.10 due to binary compatibility concerns. So we now don?t
require a value that is being unboxed to be of a final class. This should be ok since we only
unbox when a primitive argument type is expected which (we think) should only happen if
the argument is an AnyVal.