Change project structure to build and publish core library and extras separately
Kiama is now three sub-projects: base with macros and basic functionality, core with the rest of the core
library using the base, and extras with utility functionality such as Compiler/REPL support and
command-line configuration. The first two (base and core) are published together as kiama and the latter
is published as kiama-extras.