Extend report interface to support clearing Monto products
Needed to have a way to clear target-type products in a compiler if
there are syntax errors *and* if there are semantic errors, but can't
build that in to Compiler since some of them don't even have target
products. Now you override the report method to clear the products if
you have them.

- add source argument to report
- minijava: override report to also clear target products
- Monto publish methods default to empty contents