Simplify enabling higherKinds
Previously we used an import to enable higherKinds in some files.
As of 2.13.1 this is not working well because the import is not
necessary, so we got an unused import warning (and hence a
build error). Ignoring the warning is possible in 2.13.2 onwards
using @nowarn, but that causes complications for older versions
that don't have that support.

To avoid the issue in a simple way, we now turn on higherKinds
via a compiler option (and hence for all files).